requirements.txt
fastapi
uvicorn
pymongo
requests
PyMuPDF
python-multipart
bcrypt
pyjwt
aiofiles
